Major Refactor & Fixed threading measurement issues & some code cleanup - renamed main.py to FreeGrid.py - removed deprecated deconv() function - removed deprecated grid_creation module - moved most modules to a single folder named ‚components‘ - added dsp_helper module: moved deconvolution, headphone compensation and sweep generation - updated the wave files: made the samples shorter to speed up automatic measurement - measurement_trigger can now only be set when no measurement is running - Headphone deconvolution: changed the bandpass low frequency from 200 to 20Hz - removed the classes for measurement threads. The threaded HRIR measurements are started directly with a thread object called from the main callback. The center measurements have been made non-threaded - done_hrir_measurement() is no hrir_measurement() and also starts the single_measurement() - removed measurement.getIRs(). Doconvolution is now handled outside of measurement module - removed measurement.getRecordings(). Recorded data is directly returned with calling single_measurement() - re-ordered hrir_measurements() (formerly done_hrir_measurements()): 	- roughly seperated the procedure in two sections: 		- first do everything that is not thread safe. No other measurement may be triggered during this time 		- then, after measurement_running flag has been released, do time-costly things that only rely on function-local variables and thus can be executed in parallel. This section is surrounded by a simple lock. 	- introduced a measurement_tolerance_check flag that is immediately released after sweep playback, 		so a subject cannot interrupt a measurement after sweep has stopped. So measurement_running 		flag only indicated that a new measurement should not be triggered, but the angular variance c		heck is controlled via measurement_tolerance_check. 	- disabled IR plotting of HRIRs 	- moved plotting of recordings before releasing measurement_running flag 	- measurement_running flag is released AFTER success sound playback 	- checked that the doconvolution and saving to file are thread-safe and also work if multiple measurements are waiting to be processed

+++ GUI/main_window.py
@@ -773,9 +773,9 @@ class UiMainWindow(QtWidgets.QMainWindow):
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.tab_hpc), _translate("MainWindow", "Headphone Compensation"))
 
 
-        self.positions_table.setColumnWidth(0, self.positions_table.width() / 3)
-        self.positions_table.setColumnWidth(1, self.positions_table.width() / 3)
-        self.positions_table.setColumnWidth(2, self.positions_table.width() / 3)
+        self.positions_table.setColumnWidth(0, int(self.positions_table.width() / 3))
+        self.positions_table.setColumnWidth(1, int(self.positions_table.width() / 3))
+        self.positions_table.setColumnWidth(2, int(self.positions_table.width() / 3))
 
 
 
@@ -955,7 +955,7 @@ class UiMainWindow(QtWidgets.QMainWindow):
 
 
     def trigger_center_measurement(self):
-        interval = 0.5 * 1000
+        interval = int(0.5 * 1000)
         QtCore.QTimer.singleShot(interval, self.measurement_ref.trigger_center_measurement)
 
     def trigger_point_recommendation(self):
